Add TxInBlock converter

+++ core/common/src/main/java/hera/api/model/Transaction.java
@@ -23,7 +23,7 @@ public class Transaction {
 
   @Getter
   @Setter
-  protected long indexInBlock;
+  protected int indexInBlock;
 
   @Getter
   @Setter
